+++ GLView/Core/GLARCHelper.h
@@ -1,13 +1,12 @@
 //
 //  ARC Helper
 //
-//  Version 1.2.2
+//  Version 1.3
 //
 //  Created by Nick Lockwood on 05/01/2012.
 //  Copyright 2012 Charcoal Design
 //
 //  Distributed under the permissive zlib license
-//  Distributed under the permissive zlib License
 //  Get the latest version from here:
 //
 //  https://gist.github.com/1563325
@@ -19,6 +18,7 @@
 #define AH_RELEASE(x) (void)(x)
 #define AH_AUTORELEASE(x) (x)
 #define AH_SUPER_DEALLOC (void)(0)
+#define __AH_BRIDGE __bridge
 #else
 #define __AH_WEAK
 #define AH_WEAK assign
@@ -26,6 +26,7 @@
 #define AH_RELEASE(x) [(x) release]
 #define AH_AUTORELEASE(x) [(x) autorelease]
 #define AH_SUPER_DEALLOC [super dealloc]
+#define __AH_BRIDGE
 #endif
 #endif