+++ GLView/Images/GLImageView.h
@@ -1,8 +1,8 @@
 //
-//  GLModel.h
+//  GLImageView.h
 //
 //  GLView Project
-//  Version 1.2
+//  Version 1.2.1
 //
 //  Created by Nick Lockwood on 10/07/2011.
 //  Copyright 2011 Charcoal Design
@@ -32,16 +32,20 @@
 //  3. This notice may not be removed or altered from any source distribution.
 //
 
-
-#import <Foundation/Foundation.h>
+#import <UIKit/UIKit.h>
+#import "GLView.h"
+#import "GLImage.h"
 #import "GLARCHelper.h"
 
 
-@interface GLModel : NSObject
+@interface GLImageView : GLView
 
-+ (GLModel *)modelWithContentsOfFile:(NSString *)path;
-- (GLModel *)initWithContentsOfFile:(NSString *)path;
+@property (nonatomic, strong) GLImage *image;
+@property (nonatomic, strong) UIColor *blendColor;
+@property (nonatomic, copy) NSArray *animationImages;
+@property (nonatomic, assign) NSTimeInterval animationDuration;
+@property (nonatomic, assign) NSInteger animationRepeatCount;
 
-- (void)draw;
+- (GLImageView *)initWithImage:(GLImage *)image;
 
 @end